WebhookScripts/Receiver/requirements.txt
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.2
